pyproject.toml
[project]
name = "simple485-remastered"
version = "0.1.0"
description = "A simple, robust RS485 communication protocol library based on simple485."
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
]
[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
]
